PRUEBA-RPA-main/requirements.txt
# Configuración y variables de entorno
python-dotenv==1.0.0

# API y HTTP
requests==2.31.0
urllib3==2.0.4
httpx==0.24.1

# Manipulación de datos
pandas==2.0.3

# Excel
openpyxl==3.1.2
xlsxwriter==3.1.2

# Automatización web
selenium==4.15.2
webdriver-manager==4.0.1

# Microsoft Graph API (OneDrive)
msal==1.24.0
azure-identity==1.14.0

# Base de datos (SQLite viene incluido con Python)

# Utilidades de fecha y validación
python-dateutil==2.8.2
pytz==2023.3
pathvalidate==3.2.0

# Logging y monitoreo
colorlog==6.7.0

# Validación y schemas
jsonschema==4.19.0

# Procesamiento de imágenes (para screenshots)
Pillow==10.0.0

# JSON y YAML
pyyaml==6.0.1

# Retry y backoff
tenacity==8.2.3

# CLI y argumentos
click==8.1.7

# Servidor web para formulario real
Flask==3.0.3
Werkzeug==3.0.3
